+++ src/listeners/guildDelete.ts
@@ -1,7 +1,9 @@
 import { ApplyOptions } from '@sapphire/decorators';
 import { Events, Listener, ListenerOptions } from '@sapphire/framework';
 import type { Guild } from 'discord.js';
-// const lib = require('lib')({ token: process.env.STDLIB_SECRET_TOKEN });
+import leaveServerLog from '../helpers/send/leaveServerLog';
+import { AUTOCODE_ENV } from '../helpers/provide/environment';
+const lib = require('lib')({ token: process.env.STDLIB_SECRET_TOKEN });
 
 @ApplyOptions<ListenerOptions>({})
 export class UserEvent extends Listener {
@@ -16,5 +18,14 @@ export class UserEvent extends Listener {
 	public async run(guild: Guild) {
 		const guild_id = guild.id;
 		console.log(`[EVENT] ${Events.GuildDelete} - ${guild.name} (${guild_id})`);
+		await removeData(guild_id);
+
+		async function removeData(guild_id: string) {
+			await leaveServerLog(guild_id);
+			//todo: add error log if removeConfig fails
+			return await lib.chillihero['birthday-api'][AUTOCODE_ENV].guild.remove({
+				guild_id: guild_id
+			});
+		}
 	}
 }
