+++ src/listeners/guildMemberLeave.ts
@@ -1,10 +1,8 @@
 import { Events } from '@sapphire/framework';
 import { Listener } from '@sapphire/framework';
 import type { GuildMember } from 'discord.js';
-import { AUTOCODE_ENV, BOT_ID } from '../helpers/provide/environment';
+import { AUTOCODE_ENV } from '../helpers/provide/environment';
 import updateBirthdayOverview from '../helpers/update/overview';
-import leaveServerLog from '../helpers/send/leaveServerLog';
-import { UserIDEnum } from '../lib/enum/UserID.enum';
 const lib = require('lib')({ token: process.env.STDLIB_SECRET_TOKEN });
 export class UserEvent extends Listener {
 	public constructor(context: Listener.Context, options: Listener.Options) {
@@ -16,48 +14,20 @@ export class UserEvent extends Listener {
 		});
 	}
 	public async run(member: GuildMember) {
-		//TODO: remove from database and update birthday list
-		//Code: https://autocode.com/mp/chillihero/birthday-bot/dev/?filename=functions%2Fevents%2Fdiscord%2Fguild%2Fmember%2Fremove.js
-		console.log('member', member);
-
 		const user_id = member.user.id;
 		const guild_id = member.guild.id;
 
-		if (user_id === UserIDEnum.BIRTHDAYY) {
-			//todo: add custom bot server check
-			// if (customBotServer.includes(guild_id)) {
-			// 	const serverHasCustomBot = 'Dont delete anything its a custom Bot server';
-			// 	console.warn(serverHasCustomBot);
-			// 	return serverHasCustomBot;
-			// }
-			await removeData(guild_id);
-
-			//todo: add error log if removeGuildRequest fails
-		} else if (user_id === BOT_ID) {
-			// remove all Data if Bot got removed from a server
-			await removeData(guild_id);
-		} else {
-			try {
-				let removeUserRequest = await lib.chillihero['birthday-api'][AUTOCODE_ENV].birthday.delete({
-					user_id: user_id,
-					guild_id: guild_id
-				});
-				if (removeUserRequest.success) {
-					await updateBirthdayOverview(guild_id);
-				}
-			} catch (e) {
-				console.warn(`Couldn not remove birthday from left user`);
-				console.warn(e);
-			}
-		}
-		async function removeData(guild_id: string) {
-			//left user is Bot and not a custom bot server
-			await leaveServerLog(guild_id);
-			//todo: add error log if removeConfig fails
-
-			return await lib.chillihero['birthday-api'][AUTOCODE_ENV].guild.remove({
+		try {
+			let removeUserRequest = await lib.chillihero['birthday-api'][AUTOCODE_ENV].birthday.delete({
+				user_id: user_id,
 				guild_id: guild_id
 			});
+			if (removeUserRequest.success) {
+				await updateBirthdayOverview(guild_id);
+			}
+		} catch (e) {
+			console.warn(`Couldn not remove birthday from left user`);
+			console.warn(e);
 		}
 	}
 }
